Properly raise rate limit errors instead of generic Forbidden
AngelList currently expresses rate limit errors as a 403 with a JSON
object in the body containing an "over_limit" value for an "error"
attribute. Our erroneous matching of a 420 error was a legacy of
copying from the Twitter gem.

The new error class is named in anticipation of the proposed 429 HTTP
status code. We've proposed using this to the AngelList API team and
they seem keen on implementing it in the next major API version.

See: http://tools.ietf.org/html/rfc6585