Simplify header handling in error classes
Faraday responses wrap headers with a utility class that takes care of
normalizing case, converting underscored symbols, etc. -- we don't need
to do this. Furthermore, the rate limit methods could return 0 if the
headers were not present at all, which is somewhat dubious
semantically. They now return nil in this case. This is consistent with
how the Twitter gem handles these now as well.

Note the method contract change to possibly return nil here -- this
library is not really following SemVer yet, sorry :-(