Add support for /startups endpoint
Although the endpoint is documented as an endpoint for fundraising
companies, it looks more like an index for the startups which supports
a filter. Because of that I opted for calling the method #all_startups,
instead of #get_fundraising or similar, and can accept a :filter param
in the options hash. This way, if in the future this endpoint accepts
more filters or other options, we'll be covered.